api/requirements.txt
annotated-types==0.7.0
anyio==4.5.0
click==8.1.7
colorama==0.4.6
fastapi==0.115.0
greenlet==3.1.0
h11==0.14.0
idna==3.10
psycopg==3.2.2
psycopg-binary==3.2.2
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6